net: Fix TestTLSListenerCheckForInfinitLoop error check
When I run the test suite locally, this test fails with a "connection
reset by peer" error instead of EOF. Cursor tells me this is likely
timing related and happens because the connection is closed during the
TLS handshake.

We're not planning on using TLS/DTLS, but I want to be able to run the
test suite to sanity check other changes I'll be making, so fix that
here.

Now go test -v ./... passes all tests when run on my machine with go
1.24.6.

Jira: VGAPP-2477